move vacancies test to separated folder
