+++ app/vacancies/tests/__init__.py
@@ -1,0 +1,2 @@
+from .serializers import *
+from .views import *
