+++ app/vacancies/tests/views/__init__.py
@@ -1,0 +1,11 @@
+from rest_framework.test import APITestCase
+from vacancies.models import Vacancy
+from authorization.models import User
+from companies.models import Company
+from rest_framework.authtoken.models import Token
+from skills.models import Skill
+import datetime
+from decimal import *
+import mock
+
+from .vacancy import VacancyViewSetTests
